fix(orchestration): expose federated terminal resource contract
