+++ src/main/runtime/orchestration/worker-terminal-ownership.ts
@@ -49,6 +49,36 @@ export type WorkerTerminalResourceRow = {
   updated_at: string
 }
 
+export function exposeWorkerTerminalResource(resource: WorkerTerminalResourceRow): {
+  id: string
+  ownershipState: string
+  releaseState: string
+  retainedReason: string | null
+  terminalHandle: string
+  worktreeId: string | null
+  originDispatchId: string
+  ownerDispatchId: string
+  releaseRequestedAt: string | null
+  releaseCompletedAt: string | null
+  releaseError: string | null
+  archive: { source: string | null; status: string | null }
+} {
+  return {
+    id: resource.id,
+    ownershipState: resource.ownership_state,
+    releaseState: resource.release_state,
+    retainedReason: resource.retained_reason,
+    terminalHandle: resource.terminal_handle,
+    worktreeId: resource.worktree_id,
+    originDispatchId: resource.origin_dispatch_id,
+    ownerDispatchId: resource.owner_dispatch_id,
+    releaseRequestedAt: resource.release_requested_at,
+    releaseCompletedAt: resource.release_completed_at,
+    releaseError: resource.release_error,
+    archive: { source: resource.archive_source, status: resource.archive_status }
+  }
+}
+
 // Terminal state exposed by worker-list; process accounting, never Task/Dispatch outcome.
 export type WorkerTerminalListState =
   | 'active'
